+++ src/test/java/ladder/BarTest.java
@@ -2,6 +2,7 @@ package ladder;
 
 import static org.assertj.core.api.Assertions.*;
 
+import ladder.domain.Bar;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
