+++ src/test/java/ladder/LadderTest.java
@@ -2,6 +2,7 @@ package ladder;
 
 import static org.assertj.core.api.Assertions.*;
 
+import ladder.domain.Ladder;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
