+++ src/test/java/ladder/LineTest.java
@@ -1,11 +1,12 @@
 package ladder;
 
-import static ladder.Bar.IMMOVABLE;
-import static ladder.Bar.MOVABLE;
+import static ladder.domain.Bar.IMMOVABLE;
+import static ladder.domain.Bar.MOVABLE;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.ArrayList;
 import java.util.List;
+import ladder.domain.Line;
 import ladder.utils.BooleanGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
