+++ src/test/java/ladder/PlayerTest.java
@@ -2,6 +2,7 @@ package ladder;
 
 import static org.junit.jupiter.api.Assertions.*;
 
+import ladder.domain.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
